- dbjh: added install_unix.sh to match text in install.html - dbjh: updated Makefile[.unix] accordingly - dbjh: removed warning in swc.c when compiling with DJGPP - dbjh: updated install_beos.sh so that it tries to install the new ioport driver - dbjh: changed filereplace() calls to change_string() in gba.c (faster) - dbjh: coorected type in developers.html

+++ trunk/ucon64/src/backup/swc.c
@@ -106,7 +106,7 @@ int swc_write_rom(char *filename, unsigned int parport)
   }
 
   stat(filename, &fstate);
-  printf("Send: %ld Bytes (%.4f Mb)\n", fstate.st_size, (float) fstate.st_size/MBIT);
+  printf("Send: %d Bytes (%.4f Mb)\n", (int) fstate.st_size, (float) fstate.st_size/MBIT);
 
   send_command0(0xc008, 0);
   fread(buffer, 1, HEADERSIZE, file);
